requirements.txt
pillow
code128
tkcalendar
sqlalchemy